Added two new simple utility functions to use in a bash toolchain
every.py
--------
- This takes a "-n" argument and prints out a line every n lines. I find
  this nice for sampling a random number of lines from some file. Bear
in mind that its important to note the size of the file.

So for instance, a file which has 150,000 lines. If you do

cat file | every.py -n 200,000 nothing will be printed except for the
first line..

urlfind.py
----------

In work, we use a command line utility to upload git commits for review.
This prints out the url of the review page amongst other output. I
wanted a way to filter to make it just return the url.

For now, urls must begin with http(s) and although this is restrictive I
looked into implementing this using the proper rfc, but for the depth of
need in this project I ignored it.